Use two separate test suites: for Posix and for Windows
